Fix temporary CString lifetimes in CreateSoundexInfo conversion
FMOD_CREATESOUNDEXINFO was being initialized with temporary CStrings
which were freed before the function returns. This patch moves the
CStrings into the CreateSoundexInfo structure as private fields so they
will at least live as long as the original structure is not modified.
Because of this, specifying a DLS file with `dls_name` would always
fail.